menu and classnames refactor

+++ main/peripherals/DHT11.h
@@ -9,7 +9,7 @@ class Dht11 {
       sensor.begin();
     }
 
-    float getTemperature() {
+    int8_t getTemperature() {
       return sensor.readTemperature();
     }
